move function tab item
