+++ dataprep-webapp/src/app/components/suggestions-stats/actions-suggestions/_action-suggestions.scss
@@ -68,6 +68,7 @@ $search-actions-height: 25px;
 		}
 
 		.search-input + .action-suggestion-tab-items {
+			// NC:
 			height: calc(100% - #{$header-tabs-height + $search-actions-height});
 		}
 	}
