+++ dataprep-webapp/src/app/components/suggestions-stats/actions-suggestions/actions-suggestions-module.js
@@ -16,7 +16,7 @@ import uiRouter from 'angular-ui-router';
 import TALEND_WIDGET_MODULE from '../../widgets/widget-module';
 import SERVICES_STATE_MODULE from '../../../services/state/state-module';
 import SERVICES_TRANSFORMATION_MODULE from '../../../services/transformation/transformation-module';
-import TAB_ITEM_MODULE from '../tab-item/tab-item-module';
+import TAB_ITEM_MODULE from './tab-item/tab-item-module';
 
 import ActionsSuggestionsCtrl from './actions-suggestions-controller';
 import ActionsSuggestions from './actions-suggestions-directive';
