+++ dataprep-webapp/src/app/components/suggestions-stats/actions-suggestions/tab-item/tab-item-directive.js
